+++ martin/src/config/file/tiles/reload/mbtiles.rs
@@ -1,7 +1,7 @@
 use std::collections::BTreeMap;
 use std::path::PathBuf;
 
-use super::{discover_sources_by_ext, path_modified_ms};
+use crate::config::file::tiles::reload::{discover_sources_by_ext, path_modified_ms};
 
 use martin_core::tiles::{BoxedSource, mbtiles::MbtSource};
 use notify::{
